Move _generate_local_part to helpers
